+++ cvpWebApi/Models/ReportLink.cs
@@ -7,7 +7,7 @@ namespace cvpWebApi.Models
 {
     public class ReportLink
     {
-        public int report_link_id { get; set; }
+        public Int64 report_link_id { get; set; }
         public int report_id { get; set; }
         public String record_type { get; set; }
         public String report_link_no { get; set; }
